solver interface fixes

+++ src/solver/impls/ida/ida.cxx
@@ -51,7 +51,7 @@
 #define ONE         RCONST(1.0)
 
 static int idares(BoutReal t, N_Vector u, N_Vector du, N_Vector rr, void *user_data);
-static int ida_bbd_res(int Nlocal, BoutReal t, 
+static int ida_bbd_res(long Nlocal, BoutReal t, 
 		       N_Vector u, N_Vector du, N_Vector rr, void *user_data);
 static int ida_pre(BoutReal t, N_Vector yy, 	 
 		   N_Vector yp, N_Vector rr, 	 
@@ -672,7 +672,7 @@ static int idares(BoutReal t,
 }
 
 /// Residual function for BBD preconditioner
-static int ida_bbd_res(int Nlocal, BoutReal t, 
+static int ida_bbd_res(long Nlocal, BoutReal t, 
 		       N_Vector u, N_Vector du, N_Vector rr, 
 		       void *user_data)
 {
